Fix exception during first time joins (#6327)
Fixes an issue during first join with the player configuration events on
Paper. The usermap will refuse to implicitly store new users into the
usermap unless it can get a Bukkit player object for it, which is only
created during the PlayerJoinEvent. So for new users, we must defer to
then. The code that was being run (activity/mute checks) was probably
incorrect to be placed inside of configure event anyway.

Also force loads the locale bundle into memory to ensure it's ready for
servers with slow i/o.

Fixes #6315